+++ src/main/java/com/getindata/kafka/connect/iceberg/sink/converter/SinkRecordToIcebergChangeEventConverterFactory.java
@@ -1,13 +1,14 @@
 package com.getindata.kafka.connect.iceberg.sink.converter;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.getindata.kafka.connect.iceberg.sink.IcebergSinkConfiguration;
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.connect.json.JsonConverter;
 import org.apache.kafka.connect.sink.SinkRecord;
 import org.apache.kafka.connect.transforms.Transformation;
 
 public class SinkRecordToIcebergChangeEventConverterFactory {
-    public static SinkRecordToIcebergChangeEventConverter create() {
+    public static SinkRecordToIcebergChangeEventConverter create(IcebergSinkConfiguration configuration) {
         Transformation<SinkRecord> extractNewRecordStateTransformation = ExtractNewRecordStateTransformationFactory.create();
         JsonConverter keyJsonConverter = JsonConverterFactory.create(true);
         JsonConverter valueJsonConverter = JsonConverterFactory.create(false);
@@ -17,7 +18,7 @@ public class SinkRecordToIcebergChangeEventConverterFactory {
                 extractNewRecordStateTransformation,
                 keyJsonConverter,
                 valueJsonConverter,
-                keyDeserializer, valueDeserializer
-        );
+                keyDeserializer, valueDeserializer,
+                configuration);
     }
 }