+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
+- Updates AWS lakeformation transitive dependency providing lakeformation support in s3 iceberg tables.
+
 ## [Unreleased]
+
+- Updates dependencies to resolve some jackson-databind critical CVEs.
+- Updates AWS lakeformation transitive dependency providing lakeformation support in s3 iceberg tables.
+- Added Iceberg coercion support for Avro Array<Struct> types. Supports Debezium `data_collections` metadata.
+- Added support for coercion of five Debezium temporal types to their Iceberg equivalents: Date, MicroTimestamp, ZonedTimestamp, MicroTime, and ZonedTime
+- Rich temporal types are toggled on by new boolean configuration property: `rich-temporal-types`
 
 ## [0.3.1] - 2023-04-06
 
@@ -18,7 +26,6 @@
 
 -   Added support for `double` primitive type fields.
 -   Allow coercion of iceberg table identifiers to `snake_case` setting `table.snake-case` boolean configuration.
-
 ## [0.2.2] - 2023-02-17
 
 -   Allow changing iceberg-table specific settings using `iceberg.table-default.*` connector configuration properties